requirements.txt
streamlit
openai
panda
scikit-learn
python-dotenv
xgboost
imbalanced-learn
shap
matplotlib
seaborn
jupyter 
notebook
